### 2023-06-26 #593 @caropradier
eph: Caja de Herramientas para el procesamiento de la Encuesta Permanente de Hogares
Nombre de la Persona Encargada: Carolina Pradier
Usuario GitHub de la Persona Encargada: @caropradier
Usuario GitHub de las Otras Autoras del Paquete: (borra si solo hay una autora) @DiegoKoz, @pablotis,@Guidowe,@natsushok,@gefero,@jgjuara
Repositorio: https://github.com/holatam/eph
Versión Enviada:
Tipo de Envio: Estándar
Editora: @maurolepore
Revisores: @guadag12, @lidefi87

Archivo: TBD
Versión Aceptada: TBD
Idioma: es
---

-   Pega el archivo DESCRIPTION completo dentro del siguiente bloque de código.

```
Package: eph
Title: Argentina's Permanent Household Survey Data and Manipulation Utilities
Version: 0.6.0
    person("Diego", "Kozlowski", email = "diegokoz92@gmail.com", role = c("aut"), comment = c(ORCID = "0000-0002-5396-3471")),
    person("Pablo", "Tiscornia", email = "pablotisco@gmail.com", role = "aut"),
    person("Guido", "Weksler", email = "guidowe45@gmail.com", role = c("aut")),
    person("Natsumi", "Shokida", email = "natsumi.shokida@gmail.com", role = "aut"),
    person("German", "Rosati", email = "german.rosati@gmail.com", role = "aut", comment = c(ORCID = "0000-0002-9775-0435")),
    person("Juan Gabriel", "Juara", email = "jg.juara@gmail.com", role = "ctb"),
    person("Carolina", "Pradier", email = "carolinapradier@gmail.com", role = c("aut","cre"), comment = c(ORCID = "0009-0007-5058-6352")))
Description: Tools to download and manipulate the Permanent Household Survey from Argentina
    (EPH is the Spanish acronym for Permanent Household Survey).
    e.g: get_microdata() for downloading the datasets, get_poverty_lines() for downloading the official poverty baskets,
    calculate_poverty() for the calculation of stating if a household is in poverty or not, following the official methodology.
    organize_panels() is used to concatenate observations from different periods, and organize_labels()
    adds the official labels to the data. The implemented methods are based on INDEC (2016) <http://www.estadistica.ec.gba.gov.ar/dpe/images/SOCIEDAD/EPH_metodologia_22_pobreza.pdf>.
    As this package works with the argentinian Permanent Household Survey and its main audience is from this country,
    the documentation was written in Spanish.
    expss,
    readxl,
    attempt,
    zoo,
    leaflet,
    htmltools,
    rvest,
    xml2,
    cli
Language: es
    forcats
```

## Alcance

- Por favor, indica qué categoría(s) aplican a este paquete. Las puedes encontrar en nuestras [políticas de inclusión de paquetes (Inglés)](https://ropensci.github.io/dev_guide/policies.html#package-categories). Por favor, tilda todas las apropiadas. Si no estás seguro, te sugerimos que comiences un pre-envío.

	- [x] recuperación de datos (data retrieval)
	- [ ] extracción de datos (data extraction)
	- [ ] munging de datos (data munging)
	- [ ] disposición o declaración de datos (data deposition)
    - [ ] validación y prueba de datos (data validation and testing)
	- [ ] automatización de flujos de trabajo (workflow automation)
	- [ ] control de versiones (version control)
	- [ ] manejo de citas y bibliometría (citation management and bibliometrics)
	- [ ] envoltorios de software científico (scientific software wrappers)
	- [ ] herramientas para trabajo de campo y reproducibilidad (field and lab reproducibility tools)
	- [ ] ligamientos con software de base de datos (database software bindings)
	- [ ] datos geoespaciales (geospatial data)
	- [ ] análisis de texto (text analysis)

- Explica cómo y por qué el paquete encaja dentro de estas categorías (1 a 3 oraciones):

La librería eph tiene como objetivo facilitar el trabajo con las bases de datos vinculadas a la [Encuesta Permanente de Hogares - INDEC](https://www.indec.gob.ar/bases-de-datos.asp). Incluye herramientas para descargar las bases de datos, etiquetar y recodificar sus variables y acceder a datasets complementarios.

- ¿Cuál es la audiencia esperada y las aplicaciones científicas de este paquete?

Científicos sociales y otros usuarios de la EPH que utilizan sus datos para elaborar investigaciones científicas, informes de coyuntura, recomendaciones de política pública, etc.

- ¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro [criterio del mejor de su categoría (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#overlap)?

Hay otros paquetes de R que facilitan la descarga y procesamiento de otras bases de datos (e.j. [ech](https://cran.r-project.org/web/packages/ech/index.html), [PNADclBGE](https://cran.r-project.org/web/packages/PNADcIBGE/index.html), [readabs](https://cran.r-project.org/web/packages/readabs/index.html)), pero ninguno de ellos es aplicable a la EPH.

-   (Si aplica) ¿Tu paquete cumple con nuestras [guías de _Ética, Privacidad de Datos e Investigación de Sujetos Humanos_ (documento en Inglés)](https://devguide.ropensci.org/policies.html#ethics-data-privacy-and-human-subjects-research)?

Si

- Si ya has hecho una consulta de pre-envío, por favor pega el enlace al _issue_ correspondiente, una publicación del foro, u otra discusión. Alternativamente, etiqueta al editor (con @tag) con el que te contactaste.

Nota del editor (@maurolepore): Este paquete es parte del ["Champions Program"](https://ropensci.org/champions/). @yabellini esta al tanto de su envio. Yo tambien soy parte del programa y el potencial conflicto de interes fue discutido y aprobado por @noamross.

Editor's note (@maurolepore). This package is part of the ["Champions Program"](https://ropensci.org/champions/). @yabellini is aware of this submission. I am also part of the program and the potential conflict of interest was discusses and approved by @noamross.

-   (Si aplica) Explique las razones de los [elementos `pkgcheck`](https://docs.ropensci.org/pkgcheck/) que su paquete no puede pasar.

## Revisiones Técnicas

Tilda los siguientes items para confirmar que los has completado:

- [x] He leído la [guía sobre paquetes de rOpenSci](https://devguide.ropensci.org/building.html).
- [x] He leído la [guía para autores](https://devdevguide.netlify.app/authors-guide.html) y pienso mantener este paquete durante al menos 2 años o encontrar un reemplazo.

Este paquete:

- [x] No viola los Términos de Servicio de ningún servicio con los que interactúa.
- [x] Tiene una licencia aceptada por CRAN y OSI.
- [x] Contiene un [archivo README con instrucciones para instalar la versión de desarrollo](https://ropensci.github.io/dev_guide/building.html#readme).
- [x] Incluye [documentación con ejemplos para todas las funciones, creada con roxygen2](https://ropensci.github.io/dev_guide/building.html#documentation).
- [x] Contiene una viñeta (_vignette_) con ejemplos de sus funciones esenciales y su uso.
- [x] Tiene una [suite de tests (documento en Inglés)](https://devguide.ropensci.org/building.html#testing).
- [x] Tiene [integración contínua (documento en Inglés)](https://devguide.ropensci.org/ci.html), incluyendo reporte de cobertura de tests.

## Opciones de Publicación

- [x] ¿Tienes intenciones de subir este paquete a CRAN?
- [ ] ¿Tienes intenciones de enviar este paquete a Bioconductor?

- [ ] ¿Deseas enviar un Artículo de Aplicaciones sobre tu paquete a [Methods in Ecology and Evolution (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/)? Si es así:

<summary>Opciones para MEE</summary>

- [ ] Este paquete es novedoso y será de interés para la mayoría de lectores de la revista.
- [ ] El manuscrito que describe el paquete no tiene más de 3000 palabras y está escrito en Inglés.
- [ ] Tienes intenciones de archivar el código del paquete en un repositorio a largo plazo, que cumple los requerimientos de la revista (mira las [Políticas de Publicación de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/journal-resources/policy-on-publishing-code.html))
- (*Alcance: Considera los [Objetivos y Alcance de MEE (documento en Inglés)](http://besjournals.onlinelibrary.wiley.com/hub/journal/10.1111/(ISSN)2041-210X/aims-and-scope/read-full-aims-and-scope.html) para tu manuscrito. No otorgamos garatías de que tu manuscrito esté en el ámbito de MEE.*)
- (*Aunque no es requerido, recomendamos tener un manuscrito completamente preparado y en Inglés, al momento de enviar.*)
- (*Por favor, no envíes tu paquete de forma separada a Methods in Ecology and Evolution*)

## Código de Conducta

- [x] Estoy de acuerdo en cumplir el [Código de Conducta de rOpenSci (documento en Inglés)](https://ropensci.github.io/dev_guide/policies.html#code-of-conduct) durante el proceso de revisión, y en mantener mi paquete si éste es aceptado.
