### 2023-07-01 #593 @maurolepore
Gracias @caropradier por compartir tu trabajo con nosotros. Es un gusto ser el editor.

Para ayudarte a rastrear mis comentarios, los etiqueto con "ml" y una secuencia numérica: ml01, ml02, etc:

* Los comentarios iniciados con "bullet" (como este) sugieren cosas que no requieren tu acción o respuesta.

- [ ] Los comentarios iniciados con "checkbox"  (como este) piden cosas que requieren tu acción y respuesta.

Aca van mis primeros comentarios:

- [ ] ml01. ¿Confirmas que la revisión es en Castellano?

- [ ] ml02. Por favor sugerí 3 revisores/as. Aunque usaremos no mas que uno, tu sugerencia sirve para ver que tipo de experiencia sería útil^1.

Estoy por comenzar mis chequeos iniciales^2. Pronto los voy a publicar aca.

----

> ^1 The potential suggestions made by the submitter(s), (although submitters may have a narrow view of the types of expertise needed. We suggest not using more than one of suggested reviewers)
-- https://devguide.ropensci.org/editorguide.html

> ^2 After automatic checks are posted, use the [editor template](https://devguide.ropensci.org/editortemplate.html#editortemplate) to guide initial checks and record your response to the submission
-- https://devguide.ropensci.org/editorguide.html
