### 2023-07-03 #593 @caropradier
Gracias Mauro!

ml01. Confirmo que la revisión es en Castellano

ml02. Sugiero a:
1. [Karina Bartolomé](https://github.com/karbartolome)
2. [Guadalupe Gonzalez](https://github.com/guadag12)
3. [Andrea Gomez Vargas](https://github.com/SoyAndrea)
