### 2023-07-07 #593 @maurolepore
@caropradier,

Muchas gracias por tu paciencia. Aca comparto mis comentarios. Seria genial abordar los requerimientos antes de bucar revisores. Si tenes alguna duda no dudes en consultarme.

(Perdon por los horrores de ortografia)

    - [ ] Is the case for the package well made

    - [ ] ml03. Por favor argumenta la importancia y necesidad del paquete. Solo veo el objetivo:

    > "eph tiene por objecto facilitar el trabajo de aquellos usuarios y usuarias de la Encuesta Permanente de Hogares - INDEC".

    Si.

    Si.

Si.

Si.

- [ ] ml04. Por favor explica como eph cumple este requerimiento.

Si. No veo detalles sobre tokens. Asumo que son innecesarios.

Si.

Si. Veo que las "issues" pendientes son pocas.

---

##### DOCUMENTACION

###### https://holatam.github.io/eph/

- [ ] ml05. En la seccion "Modo de uso", por favor mostra los resultados asi se ven sin necesidad de installar el paquete y correr el codigo.

- [ ] ml06. Por favor usa un estilizador de codigo automatico (e.j. `usethis::use_tidy_style()`). Es util ver que el estilo es homogeneo, y que sigue una guia de estilo popular como la "tidyverse style guide" (https://style.tidyverse.org/). Eso ayuda a no pensar en el estilo y guardar energia mental para notar cosas mas importantes. Estos son algunos ejemplos de violaciones de estilo:

    * `dot.case` en vez de `snake_case`.
    * `}else{` en vez de `} else {`.
    * `#'Roxyen comments` en vez de  `#' Roxygen comments`.
    * `return(output)` en vez de `output`.

* ml07. Al escribir documentacion, considera dirigirte a una persona especifica (e.j. como hago yo en esta serie de comentarios). Cuanto te escribo a "vos" en vez de a "los usuarios y usuarias" evito referencias de genero, y "hablo" mas directa y naturalmente. Por ejemplo:

    * Original: La librería eph tiene por objecto facilitar el trabajo de aquellos usuarios y usuarias de la Encuesta Permanente de Hogares - INDEC ...

    * Revision: Si usas la Encuesta Permanente de Hogres - INDEC, la libreria eph es para vos ...

* ml08. El logo ocupa toda la pantalla. Considera achicarlo.

* ml09. La seccion "Como citar este paquete" aparece cerca del principio. Considera moverla cerca del final. Cuando exploras un nuevo paquete, probablemente queres saber como usar el paquete antes de saber como citarlo.

* ml10. Considera usar la "tidyverse style guide" tambien para los comentarios. Segun esa guia: "Each line of a comment should begin with the comment symbol and a single space" (https://style.tidyverse.org/syntax.html?q=comm#comments).

```r
# Bien
# Comentario
x <- 1

# Mal
x <- 1  # Comentario
```

Esto ayuda a mantener los comentarios, porque el espacio disponble en una linea de comentario es totalmente independiente del especio disponible en una linea para codigo.

###### https://holatam.github.io/eph/articles/eph.html

* ml11. Los nombres (German Rosati, Diego Kozlowski, etc.) ocupan mucho espacio vertical. Considera listarlos horizontalmente. En _pkgdown.yaml considera vincular cada nombre a un URL, e.j.:

  Mauro Lepore
    href: https://github.com/maurolepore
```

* ml12. Considera no duplicar la definicion de los argumentos. La duplicacion hace la documentacion dificil de manetner. Cuando cambias algo en un lugar, tenes que recordar cambiarlo en el otro lugar tambien. Por ejemplo, en `get_microdata()` la definicion del argumento `year` esta duplicada:

    * [`year` un integer o vector de integers a partir de 2003](https://holatam.github.io/eph/reference/get_microdata.html#arguments-1).
    * [`year` un valor numérico entero con el año de la encuesta que se desea descargar](https://holatam.github.io/eph/articles/eph.html).

* ml13. Considera usar `janitor::clean_names()` para convertir el nombre de las columnas a `snake_case` (consistentemente con la "tidyverse style guide").

* ml14. Cuando usas "tibbles" considera no usar `head()`. Por defecto, los tibbles muestran solo 10 filas del "data frame".

* ml15. Cuando llamas una function, evita usar el nombre del primier argumento (https://design.tidyverse.org/call-data-details.html#whats-the-pattern).

###### https://holatam.github.io/eph/articles/estimacion_pobreza.html

- [ ] ml16. Al assignar valores a objectos, usa consitentemente el simbolo `<-`, no `=` (o viceversa).

* ml17. Consiera mover la llamada a `knitr::opts_chunk$set()` a un "chunk" de codigo con `echo = FALSE`.

###### https://holatam.github.io/eph/news/index.html

* ml18. Considera referirte a `calculate_tabulats()` no calculate_tabulats() (usando el estido de `codigo`).

##### TEMPLATE

¿Hay otros paquetes de R que logren el mismo objetivo? Si los hay, ¿cómo se diferencian del tuyo, o alcanzan nuestro criterio del mejor de su categoría (documento en Inglés)?

- [ ] ml19. Por favor listar algunos ememplos. Cuando revisas un paquete es util ver otros paquetes similares.

 ¿Tienes intenciones de subir este paquete a CRAN?

- [ ] ml20. Veo que el paquete ya esta en CRAN y que vos tenes el rol de "maintainer". Estas al tando de esta politica de rOpenSci?:

> We strongly suggest submitting your package for review before publishing on CRAN or submitting a software paper describing the package to a journal. Review feedback may result in major improvements and updates to your package, including renaming and breaking changes to functions. We do not consider previous publication on CRAN or in other venues sufficient reason to not adopt reviewer or editor recommendations.

https://devguide.ropensci.org/policies.html#publishing-in-other-venues

##### ISSUES

* ml21. Las "issues" pendientes parecen antiguas (12 meses o mas). Reconsidera su vigencia.

##### CHECKS

* ml22. En "Package dependencies" veo muchos paquetes con `ncall = NA`. Verifica que aun sean utiles.

* ml23. Nota los resultados de goodpractice. Por ejemplo, se detectaron muchas lineas de codigo con mas de 80 caracteres de largo (estandard). Eso complica la revision.

##### CODE OVERVIEW

* ml24. Considera usar el suporte que roxygen ofrece para escribir Markdown: Por ejemplo, podes escribir `* ...` en vez de `\item{...}` y `aglomerado` en vez de `\code{aglomerado}`.
