+++ README.md
@@ -1,95 +1,82 @@
 
-# eph <a><img src='man/figures/logo.png' align="right" height="250" /></a>
+<!-- README.md is generated from README.Rmd. Please edit that file -->
+
+# eph <img src='man/figures/logo.png' align="right" height="200" style="float:right; height:200px;" />
 
 <!-- badges: start -->
-[![R build status](https://github.com/holatam/eph/workflows/R-CMD-check/badge.svg)](https://github.com/holatam/eph/actions)
+
+[![R build
+status](https://github.com/holatam/eph/workflows/R-CMD-check/badge.svg)](https://github.com/holatam/eph/actions)
 [![Coverage
 status](https://codecov.io/gh/holatam/eph/branch/master/graph/badge.svg)](https://app.codecov.io/gh/holatam/eph?branch=master)
-[![CRAN status](https://www.r-pkg.org/badges/version/eph)](https://cran.r-project.org/package=eph)
+[![CRAN
+status](https://www.r-pkg.org/badges/version/eph)](https://cran.r-project.org/package=eph)
 [![](http://cranlogs.r-pkg.org/badges/grand-total/eph?color=blue)](https://cran.r-project.org/package=eph)
 [![](http://cranlogs.r-pkg.org/badges/last-month/eph?color=blue)](https://cran.r-project.org/package=eph)
 [![DOI](https://zenodo.org/badge/142294444.svg)](https://zenodo.org/badge/latestdoi/142294444)
-
 
 <!-- badges: end -->
 
 # Caja de Herramientas para el procesamiento de la Encuesta Permanente de Hogares
 
 ## Descripción
-La librería `eph` tiene por objeto facilitar el trabajo de aquellos usuarios y usuarias de la [Encuesta Permanente de Hogares - INDEC](https://www.indec.gob.ar/bases-de-datos.asp) que deseen procesar datos de la misma mediante el lenguaje de programación [R](https://www.r-project.org/).
 
+La librería `eph` tiene por objeto facilitar el trabajo de aquellos
+usuarios y usuarias de la [Encuesta Permanente de Hogares -
+INDEC](https://www.indec.gob.ar/bases-de-datos.asp) que deseen procesar
+datos de la misma mediante el lenguaje de programación
+[R](https://www.r-project.org/).
 
 Algunas de sus funciones son:
 
 - **`get_microdata()`**: Descarga las bases de microdatos,
 
-- **`organize_panels()`**: Permite armar un pool de datos en panel de la EPH continua,
+- **`organize_panels()`**: Permite armar un pool de datos en panel de la
+  EPH continua,
 
 - **`organize_cno()`**: Clasifica las ocupaciones según el CNO 2001
 
-- **`organize_caes()`**: Clasifica las actividades económicas según CAES Mercosur 1.0 y CAES Mercosur
+- **`organize_caes()`**: Clasifica las actividades económicas según CAES
+  Mercosur 1.0 y CAES Mercosur
 
-- **`organize_labels()`**: Etiqueta las bases siguiendo el último [diseño de registro](https://www.indec.gob.ar/ftp/cuadros/menusuperior/eph/EPH_registro_t218.pdf)
+- **`organize_labels()`**: Etiqueta las bases siguiendo el último
+  [diseño de
+  registro](https://www.indec.gob.ar/ftp/cuadros/menusuperior/eph/EPH_registro_t218.pdf)
 
 - **`map_agglomerates()`**: Mapa de indicadores por aglomerado
 
-
-El paquete también cuenta con dataframes útiles para el trabajo con la EPH. Algunos de estos son:
+El paquete también cuenta con dataframes útiles para el trabajo con la
+EPH. Algunos de estos son:
 
 - **`diccionario_regiones`**
[…]
