+++ R/is_in_github.R
@@ -14,18 +14,21 @@ is_in_github <- function(year = 2018,
     result <- TRUE # esto es porque todas las bases de la EPH puntual estan en github
   }else{
 
-    pg <- xml2::read_html(sprintf('https://github.com/holatam/data/tree/master/eph/%s',type))
-    # req <- httr::GET("https://api.github.com/repos/holatam/data/git/trees/master?recursive=1")
+    #pg <- xml2::read_html(sprintf('https://github.com/holatam/data/tree/master/eph/%s',type))
+     req <- httr::GET("https://api.github.com/repos/holatam/data/git/trees/master?recursive=1")
 
-    filelist <- rvest::html_nodes(pg, "a") %>%
-      rvest::html_attr(name = "href" ) %>%
-      stringr::str_match('.*RDS') %>%
-      stats::na.omit()
+    # filelist <- rvest::html_nodes(pg, "a") %>%
+    #   rvest::html_attr(name = "href" ) %>%
+    #   stringr::str_match('.*RDS') %>%
+    #   stats::na.omit()
 
-    #filelist <- unlist(lapply(httr::content(req)$tree, "[", "path"), use.names = F)
-    looking_for <-sprintf('/holatam/data/blob/master/eph/%s/base_%s_%sT%s.RDS',type,type,year,period)
+    filelist <- unlist(lapply(httr::content(req)$tree, "[", "path"), use.names = F)
 
-    result <- looking_for %in% filelist[,1]
+    #looking_for <-sprintf('/holatam/data/blob/master/eph/%s/base_%s_%sT%s.RDS',type,type,year,period)
+    looking_for <-sprintf('eph/%s/base_%s_%sT%s.RDS',type,type,year,period)
+
+    #result <- looking_for %in% filelist[,1]
+    result <- looking_for %in% filelist
   }
 
   result
