### 2023-07-09 f128edf
add state of the art tests
