+++ man/errores_muestrales.Rd
@@ -7,14 +7,14 @@
 \format{
 A data frame with 1687 rows and 5 variables:
 \describe{
-  \item{\code{codigo}}{character ---String con código numérico de los 31 aglomerados, "Gran Buenos Aires"
+  * `codigo`{character ---String con código numérico de los 31 aglomerados, "Gran Buenos Aires"
    (sólo para 2003.03 a 2014.02), o con "Total" para el conjunto de los 31 aglomerados---}
-  \item{\code{aglomerado}}{character ---String con el nombre del aglomerado---}
-  \item{\code{periodo}}{character ---String indicando el periodo de EPH que corresponde, "2014.03"
+  * `aglomerado`{character ---String con el nombre del aglomerado---}
+  * `periodo`{character ---String indicando el periodo de EPH que corresponde, "2014.03"
    para datos de EPH a partir del tercer trimestre 2014, o "2003.03_2014.02" para datos anteriores---}
-  \item{\code{x}}{double ---Estimación de población para la cual se desea conocer el error muestral---}
-  \item{\code{ds}}{double ---Desvío Estándar correspondiente a la estimación de población en el aglomerado---}
-  \item{\code{cv}}{double ---Coeficiente de Variación correspondiente a la estimación de población en el aglomerado---}
+  * `x`{double ---Estimación de población para la cual se desea conocer el error muestral---}
+  *`ds`{double ---Desvío Estándar correspondiente a la estimación de población en el aglomerado---}
+  * `cv`{double ---Coeficiente de Variación correspondiente a la estimación de población en el aglomerado---}
 }
 }
 \usage{
