+++ R/diccionario_regiones.R
@@ -1,6 +1,6 @@
 #'Diccionario de regiones segun diseño de registro de la EPH
 #'
-#' @format  A data frame with 6 rows and 2 variables
+#' @format  Un data frame con 6 filas y 2 variables
 #'
 #'\describe{
 #' \item{codigo}{Codigo de region}
