+++ R/errores_muestrales.R
@@ -3,7 +3,7 @@
 #'  para la EPH continua desde 2003 segundo trimestre según documentación de INDEC:
 #' https://www.indec.gob.ar/ftp/cuadros/menusuperior/eph/EPH_errores_muestreo_3t2014.pdf
 #' https://www.indec.gob.ar/ftp/cuadros/menusuperior/eph/EPH_errores_muestreo.pdf
-#' @format A data frame with 1687 rows and 5 variables:
+#' @format Un data frame con 1687 filas y 5 variables:
 #' \describe{
 #'   * `codigo`{character ---String con código numérico de los 31 aglomerados, "Gran Buenos Aires"
 #'    (sólo para 2003.03 a 2014.02), o con "Total" para el conjunto de los 31 aglomerados---}
