+++ R/diccionario_aglomerados.R
@@ -1,6 +1,6 @@
 #'Diccionario de aglomerados segun diseño de registro de la EPH
 #'
-#' @format  A data frame with 32 rows and 2 variables
+#' @format  Un data frame con 32 filas y 2 variables
 #'
 #'\describe{
 #' \item{codigo}{Codigo de aglomerado}
