+++ man/diccionario_regiones.Rd
@@ -5,7 +5,7 @@
 \alias{diccionario_regiones}
 \title{Diccionario de regiones segun diseño de registro de la EPH}
 \format{
-A data frame with 6 rows and 2 variables
+Un data frame con 6 filas y 2 variables
 
 \describe{
 \item{codigo}{Codigo de region}
