+++ man/get_poverty_lines.Rd
@@ -18,7 +18,7 @@ Devuelve una tabla con la CBA y CBT a partir de abril 2016
 Funcion que descarga la CBA y CBT a partir de abril 2016
 }
 \details{
-Disclaimer: El script no es un producto oficial de INDEC.
+disclaimer: El script no es un producto oficial de INDEC.
 }
 \examples{
 
