+++ man/organize_labels.Rd
@@ -18,7 +18,7 @@ Devuelve la base con sus variables etiquetadas
 Funcion para etiquetar las bases de la Encuesta Permanente de Hogares.
 }
 \details{
-Disclaimer: El script no es un producto oficial de INDEC.
+disclaimer: El script no es un producto oficial de INDEC.
 }
 \examples{
 df <- organize_labels(toybase_individual_2016_04, type='individual')
