+++ man/errores_muestrales.Rd
@@ -5,7 +5,7 @@
 \alias{errores_muestrales}
 \title{Tabla con los errores muestrales para estimaciones de población}
 \format{
-A data frame with 1687 rows and 5 variables:
+Un data frame con 1687 filas y 5 variables:
 \describe{
   * `codigo`{character ---String con código numérico de los 31 aglomerados, "Gran Buenos Aires"
    (sólo para 2003.03 a 2014.02), o con "Total" para el conjunto de los 31 aglomerados---}
