+++ man/adulto_equivalente.Rd
@@ -5,7 +5,7 @@
 \alias{adulto_equivalente}
 \title{Tabla de valores de adulto equivalente segun sexo y edad}
 \format{
-A data frame with 222 rows and 3 variables
+Un data frame con 222 filas y 3 variables
 
 \describe{
 \item{CH04}{sexo}
