+++ R/organize_panels.R
@@ -1,70 +1,75 @@
 #'  Pool de Datos en Panel - Base Individudal EPH continua
 #'
-#'Permite armar un pool de datos en panel de la EPH continua a partir
-#'de especificar una serie consecutiva de bases, variables y el largo de la ventana -window- de observación
+#' Permite armar un pool de datos en panel de la EPH continua a partir
+#' de especificar una serie consecutiva de bases, variables y el largo de la ventana -window- de observación
 #'
 #' @param bases Lista de bases de microdatos a utilizar para armar el pool de datos
 #' @param variables Vector con nombres de las variables de interes
 #' @param window Especificar distancia temporal entre las observaciones. anual o trimestral
 #' @details
-#'disclaimer: El script no es un producto oficial de INDEC.
+#' disclaimer: El script no es un producto oficial de INDEC.
 #' @return Devuelve el pool de datos de panel
 #' @export
 #'
 #' @examples
 #'
-#' lista_bases <- list(toybase_individual_2016_03,toybase_individual_2016_04)
-#' pool_trimestral <- organize_panels(bases = lista_bases,
-#'                     variables =c("P21","ESTADO"),
-#'                     window = "trimestral")
-
-organize_panels <- function(bases,variables,window = "anual"){
-
-
-  if(!is.list(bases)){
+#' lista_bases <- list(toybase_individual_2016_03, toybase_individual_2016_04)
+#' pool_trimestral <- organize_panels(
+#'   bases = lista_bases,
+#'   variables = c("P21", "ESTADO"),
+#'   window = "trimestral"
+#' )
+organize_panels <- function(bases, variables, window = "anual") {
+  if (!is.list(bases)) {
     cli::cli_abort(c(
       "El argumento bases requiere un objeto de tipo lista."
     ))
   }
-  if(!window %in% c("anual", "trimestral")){
+  if (!window %in% c("anual", "trimestral")) {
     cli::cli_abort(c(
       "Elegir alguna de las opciones de window: 'anual' u 'trimestral'."
     ))
   }
 
   bases_continua <- dplyr::bind_rows(bases) %>%
-    dplyr::select(CODUSU,NRO_HOGAR,COMPONENTE,ANO4,TRIMESTRE,CH04,CH06,variables) %>%
-    dplyr::filter(ESTADO !=0) %>%
-    dplyr::mutate(Periodo = zoo::as.yearqtr(paste0(ANO4," Q",TRIMESTRE)))
+    dplyr::select(CODUSU, NRO_HOGAR, COMPONENTE,
+                  ANO4, TRIMESTRE, CH04, CH06, variables) %>%
+    dplyr::filter(ESTADO != 0) %>%
+    dplyr::mutate(Periodo = zoo::as.yearqtr(paste0(ANO4, " Q", TRIMESTRE)))
 
-  ##Creo una Replica de la base, y le agrego (_t1) al nombre de cada
-  ##variable, excepto a las que voy a usar para "matchear".
-  Teuno <-  function(x) {
-    paste0(x,"_t1")
+  ## Creo una Replica de la base, y le agrego (_t1) al nombre de cada
+  ## variable, excepto a las que voy a usar para "matchear".
+  Teuno <- function(x) {
+    paste0(x, "_t1")
   }
 
-  ##En Base a la amplitud del panel que especificaré al correr en la funcion resto en la Base
-  ##Replica el identificador de Trimestre construido, para hacer un join  con la Base.
+  ## En Base a la amplitud del panel que especificaré al correr en la funcion resto en la Base
[…]
