+++ R/canastas_reg_example.R
@@ -1,15 +1,15 @@
-#'Canastas Basicas Alimentarias y Canastas Basicas Totales segun region y trimestre
+#' Canastas Basicas Alimentarias y Canastas Basicas Totales segun region y trimestre
 #'
-#'Esta base tiene el formato necesario para el calculo de pobreza
+#' Esta base tiene el formato necesario para el calculo de pobreza
 #'
 #' @format  Un data frame con 12 filas y 5 variables
 #'
-#'\describe{
+#' \describe{
 #' \item{region}{Nombre de region}
 #' \item{periodo}{Año.trimestre}
 #' \item{CBA}{canasta basica alimentaria}
 #' \item{CBT}{inversa del coeficiente de engels}
 #' \item{codigo}{codigo de region}
-#'...
-#'}
+#' ...
+#' }
 "canastas_reg_example"

[…]
