+++ R/is_in_github.R
@@ -10,22 +10,35 @@ is_in_github <- function(year = 2018,
                          trimester = NA,
                          type='individual'){
 
-  if (is.na(trimester)) {
-    result <- TRUE # esto es porque todas las bases de la EPH puntual estan en github
-  }else{
+  if (year < 2023) {
+    result <- TRUE
+    # esto es porque todas las bases de la EPH hasta 2022 estan en github
+  } else {
 
-    pg <- xml2::read_html(glue::glue('https://github.com/holatam/data/tree/master/eph/{type}'))
-    # req <- httr::GET("https://api.github.com/repos/holatam/data/git/trees/master?recursive=1")
+    # old version #
+    # pg <- xml2::read_html(sprintf('https://github.com/holatam/data/tree/master/eph/%s',type))
 
-    filelist <- rvest::html_nodes(pg, "a") %>%
-      rvest::html_attr(name = "href" ) %>%
-      stringr::str_match('.*RDS') %>%
-      stats::na.omit()
+    # filelist <- rvest::html_nodes(pg, "a") %>%
+    #   rvest::html_attr(name = "href" ) %>%
+    #   stringr::str_match('.*RDS') %>%
+    #   stats::na.omit()
+    # looking_for <-sprintf('/holatam/data/blob/master/eph/%s/base_%s_%sT%s.RDS',type,type,year,period)
+    # result <- looking_for %in% filelist[,1]
 
-    #filelist <- unlist(lapply(httr::content(req)$tree, "[", "path"), use.names = F)
-    looking_for <-glue::glue('/holatam/data/blob/master/eph/{type}/base_{type}_{year}T{trimester}.RDS')
+    result <- tryCatch(
+      {
+        req <- httr::GET("https://api.github.com/repos/holatam/data/git/trees/master?recursive=1")
 
-    result <- looking_for %in% filelist[,1]
+        filelist <- unlist(lapply(httr::content(req)$tree, "[", "path"), use.names = F)
+
+        looking_for <- sprintf("eph/%s/base_%s_%sT%s.RDS", type, type, year, period)
+
+        result <- looking_for %in% filelist
+      },
+      error = function(err) {
+        FALSE
+      }
+    )
   }
 
   result
