+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: eph
 Title: Argentina's Permanent Household Survey Data and Manipulation Utilities
-Version: 0.6.0
+Version: 0.6.1
 Authors@R: c(
     person("Diego", "Kozlowski", email = "diegokoz92@gmail.com", role = c("aut"), comment = c(ORCID = "0000-0002-5396-3471")),
     person("Pablo", "Tiscornia", email = "pablotisco@gmail.com", role = "aut"),
@@ -37,7 +37,6 @@ Imports:
     leaflet,
     htmltools,
     rlang,
-    rvest,
     stats,
     cli,
     httr
