+++ R/get_total_urbano.R
@@ -1,4 +1,4 @@
-#' Descarga de Bases de EPH total urbano
+#' (coming soon!) Descarga de Bases de EPH total urbano
 #' @description
 #' Funcion que descarga bases de la Encuesta Permanente de Hogares total urbano del INDEC a partir de 2016
 #' @param year un integer o vector de integers a partir de 2016
