+++ R/get_eahu.R
@@ -1,4 +1,4 @@
-#' Descarga de Bases de la Encuesta anual de hogares urbanos
+#' (coming soon!) Descarga de Bases de la Encuesta anual de hogares urbanos
 #' @description
 #' Funcion que descarga bases de la Encuesta anual de hogares urbanos del INDEC entre 2010 y 2014
 #' @param year un integer o vector de integers entre 2010 y 2014
