+++ man/organize_cno.Rd
@@ -25,6 +25,6 @@ disclaimer: El script no es un producto oficial de INDEC.
 
 
 bases <- dplyr::bind_rows(toybase_individual_2016_03, toybase_individual_2016_04)
-bases_clasif <- organize_cno(base = bases)
+bases_clasificadas <- organize_cno(base = bases)
 
 }
