+++ man/caes.Rd
@@ -7,15 +7,15 @@
 \format{
 Un data frame con 140 filas y 8 variables:
 \describe{
-  * `caes_version`{character ---Version del CAES utilizada---}
-  * `PP04B_COD`{character ---Codigo clase CAES. 4 digitos (nivel 3) Variable que aparece en microdatos EPH---}
-  * `PP04B_label`{character ---Etiqueta clase CAES---}
-  * `caes_seccion_cod`{character ---Codigo seccion CAES. 2 digitos (nivel 2)---}
-  * `caes_seccion_label`{character ---Etiqueta seccion CAES---}
-  * `caes_division_cod`{character ---Codigo division CAES. Letra (nivel 1)---}
-  * `caes_division_label`{character ---Etiqueta division CAES---}
-  * `caes_eph_cod`{character ---Codigo reagrupamiento de CAES utilizado por EPH (nivel 1)---}
-  * `caes_eph_label`{factor ---Etiqueta reagrupamiento de CAES utilizado por EPH (nivel 1)---}
+  \item{caes_version}{(character) Version del CAES utilizada}
+  \item{PP04B_COD}{(character) Codigo clase CAES. 4 digitos (nivel 3) Variable que aparece en microdatos EPH}
+  \item{PP04B_label}{(character) Etiqueta clase CAES}
+  \item{caes_seccion_cod}{(character) Codigo seccion CAES. 2 digitos (nivel 2)}
+  \item{caes_seccion_label}{(character) Etiqueta seccion CAES}
+  \item{caes_division_cod}{(character) Codigo division CAES. Letra (nivel 1)}
+  \item{caes_division_label}{(character) Etiqueta division CAES}
+  \item{caes_eph_cod}{(character) Codigo reagrupamiento de CAES utilizado por EPH (nivel 1)}
+  \item{caes_eph_label}{(factor) Etiqueta reagrupamiento de CAES utilizado por EPH (nivel 1)}
 }
 }
 \usage{
