+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
+library(httptest)
 library(eph)
 
 test_check("eph")
