### 2023-07-17 #593 @caropradier
Hola Mauro! ¿Cómo estás?
Muchas gracias por tus comentarios y perdón por la demora. Estuve trabajando sobre estos comentarios y tengo algunas dudas/comentarios:

**ml04**. A los fines de cumplir con este requerimiento, agregué tests adicionales para la función interna is_in_github() para verificar el buen funcionamiento de los requests a la API de Github usando el paquete httptest.
Sin embargo, no estoy segura de si sería necesario agregar tests a la función map_agglomerates() (que utiliza el paquete leaflet) para terminar de cumplir con esta condición; en este caso, tendrías algún consejo sobre qué herramientas puedo usar para implementarlos?

**ml13**. Los nombres de las columnas se mantienen con su formato original para imitar en la medida de lo posible las bases que se descargaría el usuario manualmente si no utilizara el paquete, y así preservar la compatibilidad con la documentación que publica INDEC.

**ml15**. ¿Estabas pensando en algún caso en particular? No encontré ninguna situación en la que esto pasara, únicamente se hace mucho énfasis en el nombre de los argumentos en las funciones que ilustran su funcionamiento (con fines expositivos).

**ml19**. Comento algunos ejemplos:
https://cran.r-project.org/web/packages/ech/index.html
https://cran.r-project.org/web/packages/PNADcIBGE/index.html
https://cran.r-project.org/web/packages/readabs/index.html

**ml20**. Estoy al tanto. El objetivo es publicar una nueva versión en CRAN una vez que se complete el proceso de revisión

**ml21**. Quedan propuestas de funciones nuevas que creería que exceden el alcance del paquete pero pueden ser ideas interesantes para otro paquete, debería cerrar estos issues?

**ml22**. Todos los paquetes que aparecen listados como Imports se utilizan por lo menos una vez (se puede verificar con R-CMD check).
