### 2023-07-23 #593 @maurolepore
Gracias @caropradier,

ml04

> No estoy segura de si sería necesario agregar tests a la función `map_agglomerates()`. ... Tendrías algún consejo sobre qué herramientas puedo usar para implementarlos?

Nuestra guia recomienda testear plots con vdiffr o snapshot tests.

> For testing your functions creating plots, we suggest using vdiffr, an
> extension of the testthat package; or testthat snapshot tests.
> -- https://devguide.ropensci.org/building.html#testing

Sin embargo, los tests de aspectos visuales suelen ser sensibles a cambios
insignificantes y por lo tanto dificiles de mantener. En la medida de lo
posible, mi consejo es intentar testear el objecto detras del grafico. Podes
asignar el plot a un objecto e investigar su estructura para ver que elementos
pueden servirte para escribir los tests.

<summary>Ejemplo</summary>

``` r
library(eph)
library(testthat)

data <- data.frame(AGLOMERADO = 2L, tasa_actividad = 0.2)
p <- map_agglomerates(data, agglomerates = AGLOMERADO, indicator = tasa_actividad)
#> Assuming "lon" and "lat" are longitude and latitude, respectively
names(p)
#> [1] "x"             "width"         "height"        "sizingPolicy"
#> [5] "dependencies"  "elementId"     "preRenderHook" "jsHooks"
attributes(p$x)
#> $names
#> [1] "options" "calls"   "limits"
#>
#> $leafletData
#>   AGLOMERADO indicator nombre_aglomerado       lon       lat
#> 1          2       0.2     Gran La Plata -57.97302 -34.94983

test_that("el `AGLOMERADO` es el esperado", {
  actual <- attributes(p$x)$leafletData$AGLOMERADO
  expected <- data$AGLOMERADO
  expect_equal(actual, expected)
})
#> Test passed 🎉
```

ml13

ml15

> ¿Estabas pensando en algún caso en particular? ... Unicamente se hace mucho énfasis en
> el nombre de los argumentos en las funciones que ilustran su funcionamiento (con
> fines expositivos).

Lo note en los ejemplos de este articulo: https://holatam.github.io/eph/articles/eph.html

OK :-)

ml19

Gracias! El comentario inicial ahora incluye tus ejemplos.

ml20

:+1:

ml21

Quizas podrias crear un repo para hospedar un borrador del proximo paquete y mover ahi
las issues que sean relevantes? Asi podrias cerrarlas en eph para mantenerlo enfocado, y
aun asi asegurarte que las buenas ideas no se van a perder. Las issues quedarian vinculadas
automaticamente asi que podrias seguirles el rastro.

ml22

:+1:

----

Ademas:

- [ ] ml25 En `get_microdata()` el argumento `trimester` tiene un comportamiento
sorprendente. Quizas no deberia tener "default", o el default deberia ser un valor
que resulte en output util? (Quizas otras functiones tengan problemas similares?)

<details><summary>Ejemplo</summary>

``` r
library(eph)

# Parece que `trimester` es requerido. Quizas no deberia tener default?
# https://design.tidyverse.org/def-required.html
args(get_microdata)
#> function (year = 2018, trimester = NA, wave = NA, type = "individual",
#>     vars = "all", destfile = NA)
#> NULL
out <- get_microdata()
#> Warning in get_microdata(): No se pudo descargar la base de year 2018,trimester NA, wave NA, type individual.
#> Mensaje: (is.numeric(trimester) | is.numeric(wave)) is not TRUE
# El resultado no es un error pero tampoco parece util
dim(out)
#> [1] 0 0

# Ademas el resultado no es estable: Arriba tengo 0 columnas y aca tengo 177
# https://design.tidyverse.org/out-type-stability.html
dim(get_microdata(trimester = 1))
#> [1] 57951   177
dim(get_microdata(trimester = 2))
#> [1] 57835   177
```
