+++ man/get_eahu.Rd
@@ -4,7 +4,7 @@
 \alias{get_eahu}
 \title{(coming soon!) Descarga de Bases de la Encuesta anual de hogares urbanos}
 \usage{
-get_eahu(year, type = "individual", vars = "all", destfile = NULL)
+get_eahu(year = 2010, type = "individual", vars = "all", destfile = NULL)
 }
 \arguments{
 \item{year}{un integer o vector de integers entre 2010 y 2014}
