+++ R/get_microdata.R
@@ -59,7 +59,7 @@ get_microdata <- function(year = 2018,
 
   if (is.null(destfile) | !destfile_exists) {
     attempt::stop_if_not(
-      .x = is_online(),
+      .x = curl::has_internet(),
       msg = "No se detecto acceso a internet. Por favor checkea tu conexion."
     )
 
