+++ DESCRIPTION
@@ -34,7 +34,8 @@ Imports:
     rlang,
     stats,
     cli,
-    httr
+    httr,
+    curl
 Depends: R (>= 2.10)
 License: MIT + file LICENSE
 Encoding: UTF-8
