### 2023-08-03 #593 @maurolepore
@caropradier,

Revisando los comentarios veo que los siguientes items requeridos aparecen como pendientes y no encuentro una respuesta especifica que se refiera a ellos. Sin embargo veo en el repo que algunos ya fueron resueltos, y quizas eso aplique a todos. Podrias por favor confirmar si ya resolviste todos o justificar si no?

Creo que si exploro profundamente el repo podria encontra la respuesta pero te agradeceria si recordas el estado y podes comentar aca.

Los items a confirmar son , ml03, ml05, ml06, ml16.
