### 2023-08-03 #593 @caropradier
Hola Mauro!
Perdón, no me di cuenta de que no te había avisado. Ya están resueltos.

**ml03**. El texto actual incluye la importancia y necesidad del paquete:

Si querés procesar datos de la [Encuesta Permanente de Hogares - INDEC](https://www.indec.gob.ar/bases-de-datos.asp) mediante el lenguaje de programación [R](https://www.r-project.org/), la librería eph tiene por objeto facilitar tu trabajo.

_El paquete cumple un rol fundamental en la democratización de la posibilidad de procesar los datos publicados por INDEC y así obtener conclusiones independientes de aquellas publicadas en los informes elaborados por el organismo. Dado que la información de la EPH constituye una de las principales fuentes para el análisis de las problemáticas sociales presentes en Argentina, el paquete no sólo posibilita investigaciones académicas y periodísticas, sino que también contribuye a la formulación de políticas públicas fundamentadas en evidencia._

**ml05**. Se incluye el resultado

**ml06**. Efectivamente, apliqué usethis::use_tidy_style() al repositorio, y verifiqué que los cambios fueran acordes a las falencias que mencionabas

**ml16**. Efectivamente, actualmente se utiliza en todos los casos <-
