### 2023-09-07 #593 @guadag12
Hola, les dejo mi revisión.

Como pueden ver se centra en aspectos relacionados con la funcionalidad del paquete en cuestión y la calidad de su documentación😊.  Espero que les sea útil.

## Revisión de EPH

*Por favor trata de marcar tantas casillas como te sea posible y elabora tus argumentos en comentarios abajo de cada una. Tu revisión no esta limitada a estos temas, tal como se describe en la guia para revisores (Reviewer Guide)*

**Por favor describe cualquier relación de trabajo que tengas/hayas tenido con los autores del paquete)**

- [x] Como revisor confirmo que no tengo [conflictos de interés](https://devguide.ropensci.org/policies.html#coi) para poder hacer la revisión de este trabajo (si no estas segura si tienes un conflicto por favor entra en contacto con tu editor _antes_ de arrancar con la revisión.

Antes que nada, felicitaciones por el paquete.  Es un paquete sumamente útil para la comunidad y uno de los más importantes para aquellos que trabajan en política pública. Además de que es un *ejemplo para otros organismos del estado de cómo se puede disponibilizar data que sea en servicio de la comunidad*. Además, es una linda muestra de cómo se puede mejorar un proyecto con el tiempo y no sólo se sostiene en el mismo sino que está permanentemente creciendo.

En terminos generales funciona muy bien, corre rápido, es simple de usar y tiene mucha información.  Ahora si, procedo a la review siguiendo las [recomendaciones de rOpenSci:](https://devguide.ropensci.org/reviewtemplatees.html#reviewtemplatees "recomendaciones para la review en español de rOpenSci")

#### Documentación
- [x] **Una declaración de necesidades** que claramente describe las necesidades que el software esta diseñado a resolver y el public meta que busca atender en el archivo README
- [x] **Instrucciones de instalación** de la versión en desarrollo del paquete incluyendo cualquier dependencia no-estándar en el archivo README
- [x] **Viñeta(s)** demostrando la funcionalidad principal que ademas corren localmente. Amé la vignette y la [web](http:/holatam.github.io/eph// "web") de eph 💕
- [x] **Documentación de las funciones** exportadas
- [x] **Directrices comunitarias** incluyendo una guia de contribución en el archivo README o el archivo CONTRIBUTING y un archivo DESCRIPTION que incluye `URL`, `BugReports` and `Maintainer` (todas en inglés por convención y para que puedan ser autogeneradas con `Authors@R`)
- [ ] **Ejemplos** (que corren localmente) para todas las funciones exportadas.

Acá me gustaría hacer una aclaración. Hay ejemplos de funciones que no me corrieron localmente cuando las corrí, lanzaron error.   Me refiero a los ejemplos que figuran en los r files de github (por ejemplo [acá](https://github.com/holatam/eph/blob/master/R/get_total_urbano.R "acá"))

Son las siguientes:

1.  `get_total_urbano()`:

```r
   base_individual <- get_total_urbano(
    year = 2016:2019,
    type = "individual",
    vars = c("AGLOMERADO", "PONDERA", "ESTADO", "CAT_OCUP")
    )

#Error in get_total_urbano(year = 2016:2019, type = "individual", vars = c("AGLOMERADO",  : could not find function "get_total_urbano"
```
2.  `get_microdata()`:

```r
 base_individual <- get_microdata(
   year = 2018:2019,period = 1,
   type = "individual",
   vars = c("PONDERA", "ESTADO", "CAT_OCUP")
    )
#Error in get_microdata(year = 2018:2019, period = 1, type = "individual",  :  unused argument (period = 1)
```

3.  `get_eahu()`:
```r
 base_individual <- get_eahu(
     year = 2010:2012,
     type = "individual",
      vars = c("SUBDOMINIO", "PONDERA", "ESTADO", "CAT_OCUP")
    )
#Error in get_eahu(year = 2010:2012, type = "individual", vars = c("SUBDOMINIO",  :  could not find function "get_eahu"
```

#### Funcionalidad

- [x] **Instalación:** La instalación se completa con éxito tal como fue documentada.
- [x] **Funcionalidad:** Toda afirmación de funcionalidad del software se confirma como existente.
- [x] **Desempeño:** Toda afirmación de desempeño del software se confirma como alcanzada.

En este punto hay dos funciones cuyo elapsed time es mayor a 5 segundos. Sin embargo, me parece que siguen corriendo en un tiempo prudencial. No es un aspecto que me parezca grave, simplemente si pueden chequearlo estaría genial, pero no me parece grave.

```r
library(devtools)
mydir <- file.path (tempdir (), "srr-demo")
gert::git_clone ("https://github.com/holatam/eph", path = mydir)
devtools::check(mydir)

# checking examples (1m 53.3s)
# Examples with CPU (user + system) or elapsed time > 5s
# user system elapsed
# get_total_urbano 30.08   0.67   97.26
# get_eahu          4.11   0.22    9.33

#0 errors v | 1 warning x | 0 notes v

```

- [x] **Pruebas automáticas:** Hay pruebas unitarias que cubren las funciones esenciales dentro del paquete con un rango razonable de entradas y condiciones. Todas las pruebas corren en la maquina local.

```r
library(goodpractice)
library(eph)
pkg_path <- system.file( package = "eph")
g <- gp(pkg_path)
g
```
Hice pruebas unitarias con `testthat`
y hay algunas funciones que no tienen su test unitario en la carpeta de [tests](https://github.com/holatam/eph/tree/master/tests "tests"). Sin embargo, hay un 79% que si tienen. Como verán en el listado de abajo hay algunos que tal vez no es necesario que esten al 100% y hay otros que si. Depende mucho de cada función en particular. Les dejo el listado para que vea cuáles son y si creen necesario reforzar los test en algunas funciones por sobre otras, esta perfecto:

```r
library(covr)
cov <- package_coverage(mydir)
report(cov)
print(cov)

#eph Coverage: 78.79%
#R/get_microdata.R: 0.00%
#R/get_microdata_internal.R: 0.00%
#R/get_poverty_lines.R: 0.00%
#R/is_in_github.R: 0.00%
#R/get_total_urbano_internal.R: 11.54%
#R/get_eahu_internal.R: 16.22%
#R/calculate_tabulates.R: 53.33%
#R/calculate_errors.R: 57.14%
#R/organize_caes.R: 74.07%
#R/calculate_poverty.R: 80.95%
#R/organize_panels.R: 84.62%
#R/get_eahu.R: 87.04%
#R/get_total_urbano.R: 87.04%
#R/organize_labels.R: 99.56%
#R/map_agglomerates.R: 100.00%
#R/organize_cno.R: 100.00%

```

- [x] **Directrices de empaque**: El paquete cumple con las directrices de empaque de rOpenSci.

Si, efectivamente sigue con la directrices de empaque de rOpenSci. Teniendo en cuenta, el nombre del paquete (corto, descriptivo, haciendo eco de su unicidad). En Windows al menos funciona bien. Los nombres de las funciones siguen el esquema `object_verb()` y es compatible su uso con `Tidyverse` y los nombres de las funciones no se confunden con las de otros paquetes como `ggplot2`, `dplyr`, `magrittr`, `data.table`. Ademas, usan `message()` y `warning()` siempre que es necesario. Además el uso de `<-` sobre el `=` es consistente en todas las funciones.

- [x] ¿El código cumple con los principios generales de la **guía de revisión de Mozilla?** Sí, el código es simple, eficiente y cada función es intuitiva de entender y emplear.

Estimación de horas dedicadas a la revisión: 5 horas

- [x] Si la o las persona(s) autora(s) lo considera(n) apropiado, yo estoy de acuerdo con que me reconozcan como revisor del paquete (el rol "rev') en la el archivo DESCRIPTION del paquete.

---

### Comentarios de la revisión

Hay dos errores particulares que también me gustaría mencionar, pero que me parecen opcionales en su ejecución.

El primero es que muchas líneas de código superan los 80 caracteres. Es un error recurrente, pero si el código asi lo requiere creo que esta bien. Ademas, esta prolijo y se lee fácil.  Es un error que salta cuando hago el chequeo con `pkgcheck`, pero tambien con `goodpractice`:

```r
library(pkgcheck)
Sys.setenv ("GITHUB_TOKEN" = "<token>")
mydir <- file.path (tempdir (), "srr-demo")
gert::git_clone ("https://github.com/holatam/eph", path = mydir)
x <- pkgcheck (mydir)
print(x)
```
El segundo es que hay caracter Latin-1 en los .R files según `goodpractice` package.

```r
pkg_path <- system.file( package = "eph")
g <- gp(pkg_path)
g

#fix this R CMD check NOTE: Note: found 1 marked Latin-1 string Note: found 1508 marked UTF-8 strings
```

Cuando intento averiguar qué archivo es, me figuran los siguientes (eliminando los .Rmd y .html por supuesto)...

```r
library(readr)

# Function to search for Latin-1 strings in a file
search_encoding <- function(file_path) {
  # Guess the encoding of the file
  guessed_encoding <- readr::guess_encoding(file_path, n_max = 100)

  # Check if Latin-1 appears in the guessed encoding
  if ("ISO-8859-1" %in% guessed_encoding$encoding) {
    message(paste("Possible Latin-1 string in file:", file_path))
    print(guessed_encoding)
  }
}
# Directory of your package (change this path)
pkg_dir <- "C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo"
# List all files in the package directory recursively
all_files <- list.files(pkg_dir, pattern = "\\.(R|r|Rmd|Rnw|txt|csv)$", recursive = TRUE, full.names = TRUE)

# Search for Latin-1 strings in each file
for (file_path in all_files) {
  try({
    search_encoding(file_path)
  }, silent = TRUE)
}

# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/organize_panels.R
# PPossible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/diccionario_aglomerados.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/canastas_reg_example.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/calculate_errors.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/doc/eph.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/data-raw/caes.R
# ossible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/organize_cno.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/errores_muestrales.R
# Possible Latin-1 string in file: C:/Users/User/AppData/Local/Temp/Rtmp2jNd9Y/srr-demo/R/diccionario_regiones.R

```
Por último, para cerrar, **los quiero felicitar**🎉. Es un gran trabajo el que ha hecho y nos simplifican mucho el trabajo a todos los que trabajamos con política pública.

De todo lo que mencioné tal vez le prestaría atención a dos aspectos:
1. Los ejemplos de las funciones que mencionaba arriba. Creo que en alguna actualización quedaron desfasados y lo modificaría.
2. Los test unitarios en las funciones que consideren más relevantes los agregaría. Insisto, tal vez no es necesario que todas las funciones tengan al 100% los test, pero a los que entiendan como más relevantes si lo agregaría.

Una vez que eso está, creo que estamos para avanzar 😊. Gran trabajo,

Saludos!

Todo esto lo corrí en mi computadora que tiene la siguiente versión:
```r

# R version 4.1.1 (2021-08-10)
# Platform: x86_64-w64-mingw32/x64 (64-bit)
# Running under: Windows 10 x64 (build 19045)
#
# Matrix products: default
#
# locale:
#   [1] LC_COLLATE=Spanish_Argentina.1252  LC_CTYPE=Spanish_Argentina.1252
# [3] LC_MONETARY=Spanish_Argentina.1252 LC_NUMERIC=C
# [5] LC_TIME=Spanish_Argentina.1252
#
# attached base packages:
#   [1] stats     graphics  grDevices utils     datasets  methods   base
#
# other attached packages:
#   [1] forcats_1.0.0      ggplot2_3.4.1      lubridate_1.9.2    readr_2.1.4
# [5] tidyr_1.3.0        dplyr_1.1.2        eph_0.6.1          pkgcheck_0.1.2.003
#
# loaded via a namespace (and not attached):
#   [1] ghql_0.1.0               colorspace_2.1-0         selectr_0.4-2
# [4] ellipsis_0.3.2           class_7.3-19             gitcreds_0.1.2
# [7] leaflet_2.2.0            rprojroot_2.0.3          snakecase_0.11.0
# [10] htmlTable_2.4.1          fs_1.5.2                 httpcode_0.3.0
# [13] rstudioapi_0.15.0        proxy_0.4-27             farver_2.1.1
# [16] roxygen2_7.2.3           urltools_1.7.3           remotes_2.4.2.1
# [19] gh_1.4.0                 bit64_4.0.5              fansi_1.0.4
# [22] xml2_1.3.3               splines_4.1.1            knitr_1.43.15
# [25] jsonlite_1.8.4           whoami_1.3.0             maditr_0.8.3
# [28] compiler_4.1.1           httr_1.4.7               backports_1.4.1
# [31] Matrix_1.5-3             assertthat_0.2.1         fastmap_1.1.1
# [34] lazyeval_0.2.2           cli_3.6.0                leaflet.providers_1.13.0
# [37] visNetwork_2.1.2         htmltools_0.5.5          prettyunits_1.1.1
# [40] tools_4.1.1              igraph_1.4.2             gtable_0.3.1
# [43] glue_1.6.2               rappdirs_0.3.3           Rcpp_1.0.10
# [46] jquerylib_0.1.4          xopen_1.0.0              vctrs_0.6.1
# [49] crul_1.4.0               nlme_3.1-152             crosstalk_1.2.0
# [52] xfun_0.39                stringr_1.5.0            ps_1.7.2
# [55] brio_1.1.3               rvest_1.0.3              timechange_0.2.0
# [58] lifecycle_1.0.3          sys_3.4.1                goodpractice_1.0.4
# [61] srr_0.0.1.194            scales_1.2.1             vroom_1.6.3
# [64] clisymbols_1.2.0         hms_1.1.3                credentials_1.3.2
# [67] rex_1.2.1                parallel_4.1.1           xmlparsedata_1.0.5
# [70] RColorBrewer_1.1-3       httr2_0.2.2              gert_1.9.2
# [73] yaml_2.3.7               curl_5.0.0               gridExtra_2.3
# [76] lintr_3.1.0              rcmdcheck_1.4.0          triebeard_0.4.1
# [79] stringi_1.7.12           desc_1.4.2               e1071_1.7-13
# [82] checkmate_2.1.0          cyclocomp_1.1.1          pkgbuild_1.4.2
# [85] attempt_0.3.1            matrixStats_0.63.0       rlang_1.1.0
# [88] pkgconfig_2.0.3          lattice_0.20-44          evaluate_0.21
# [91] purrr_1.0.1              sf_1.0-14                labeling_0.4.2
# [94] htmlwidgets_1.6.1        bit_4.0.5                processx_3.8.0
# [97] tidyselect_1.2.0         magrittr_2.0.3           R6_2.5.1
# [100] expss_0.11.6             generics_0.1.3           DBI_1.1.3
# [103] mgcv_1.8-36              pillar_1.9.0             withr_2.5.0
# [106] units_0.8-1              tibble_3.2.1             praise_1.0.0
# [109] janitor_2.2.0            crayon_1.5.2             KernSmooth_2.23-20
# [112] utf8_1.2.3               tzdb_0.3.0               rmarkdown_2.24
# [115] viridis_0.6.2            usethis_2.1.6            grid_4.1.1
# [118] data.table_1.14.8        callr_3.7.3              graphql_1.5.1
# [121] digest_0.6.31            classInt_0.4-9           covr_3.6.2
# [124] openssl_2.0.6            munsell_0.5.0            viridisLite_0.4.1
# [127] askpass_1.1              pkgstats_0.1.3.008

```
