### 2023-09-09 #593 @lidefi87
---
title: "review"
  rmarkdown::md_document:
    pandoc_args: [
      "--wrap=none"
    ]
---

## Revisión de un paquete

*Por favor trata de marcar tantas casillas como te sea posible y elabora tus argumentos en comentarios abajo de cada una. Tu revisión no esta limitada a estos temas, tal como se describe en la guia para revisores (Reviewer Guide)*

**Por favor describe cualquier relación de trabajo que tengas/hayas tenido con los autores del paquete)**

- [x] Como revisor confirmo que no tengo [conflictos de interés](https://devguide.ropensci.org/policies.html#coi) para poder hacer la revisión de este trabajo (si no estas segura si tienes un conflicto por favor entra en contacto con tu editor _antes_ de arrancar con la revisión.

#### Documentación

El paquete incluye todos los siguiente tipos de documentación:

- [x] **Una declaración de necesidades** que claramente describe las necesidades que el software está diseñado a resolver y el público meta que busca atender en el archivo README
- [x] **Instrucciones de instalación** de la versión en desarrollo del paquete incluyendo cualquier dependencia no-estándar en el archivo README
- [x] **Viñeta(s)** demostrando la funcionalidad principal que además corren localmente
- [x] **Documentación de las funciones** exportadas
- [ ] **Ejemplos** (que corren localmente) para todas las funciones exportadas
- [x] **Directrices comunitarias** incluyendo una guía de contribución en el archivo README o el archivo CONTRIBUTING y un archivo DESCRIPTION que incluye `URL`, `BugReports` and `Maintainer` (todas en inglés por convención y para que puedan ser autogeneradas con `Authors@R`).

#### Funcionalidad

- [x] **Instalación:** La instalación se completa con éxito tal como fue documentada.
- [x] **Funcionalidad:** Toda afirmación de funcionalidad del software se confirma como existente.
- [x] **Desempeño:** Toda afirmación de desempeño del software se confirma como alcanzada.
- [x] **Pruebas automáticas:** Hay pruebas unitarias que cubren las funciones esenciales dentro del paquete con un rango razonable de entradas y condiciones. Todas las pruebas corren en la maquina local.
- [x] **Directrices de empaque**: El paquete cumple con las directrices de empaque de rOpenSci.

Estimación de horas dedicadas a la revisión: 5 horas

- [x] Si la o las persona(s) autora(s) lo considera(n) apropiado, yo estoy de acuerdo con que me reconozcan como revisor del paquete (el rol "rev') en la el archivo DESCRIPTION del paquete.

---

### Comentarios de la revisión
Buen trabajo creando este paquete. De seguro será de gran utilidad para quienes quieran analizar este tipo de datos complejos. Les incluyo mis comentarios abajo.

- **¿Cumple el código con los principios generales en la Guía de Revisión Mozilla?**
En general sí. Creo que las funciones son fácil de entender y en su mayoría las viñetas de las funciones detallan el objetivo de ellas. Sin embargo, creo que ciertas secciones del sitio web y algunas viñetas podrían ser mejor documentadas. He incluido más detalles bajo **Comentarios generales**.

- **¿Cumple con directrices de empaque de rOpenSci?**
En general sí, pero creo que el nombre de las funciones podrían ser un poco más estandarizadas para que sea posible reconocerlas como parte de este paquete con más facilidad. Al momento las funciones tienen nombres en español, inglés y una combinación de ambos idiomas. Quizás valga la pena considerar un prefijo también tal como lo hacen los paquetes `stringr`, `sf` o `star`. Esta es una sugerencia solamente, pero ayudaría a estar mejor alineados a la guía de rOpenSci.

**Comentarios generales**
En general la documentación del paquete es clara, pero contiene varios acrónimos que no están definidos en el texto (por ej., INDEC, CNO, CAES, incluso EPH). Creo que valdría la pena definir estos acrónimos la primera vez que son referenciados en el texto. Esto es desde el punto de vista de alguien que no vive en la Argentina.

El sitio web relacionado a este paquete es bastante detallado y los ejemplos corren de manera local sin problema. La viñetas son en gran parte informativas, pero creo que ciertas funciones deberían incluir un poco más de contexto. Por ejemplo, sugeriría que la función `organize_labels()` incluya un poco más de información acerca del "etiquetado". Inicialmente pensé que esta función resultaría en nombres de columnas más descriptivas y no encontré un cambio al imprimir el nombre de las columnas en la consola usando (`names(df)`). No fue sino hasta que vi el set de datos en una pestaña aparte que entendí qué es lo que hacía esta función. Pero debo recalcar que no acostumbro usar las pestañas para ver datos porque suelo trabajar con sets de datos bastante grandes. En las viñetas, `adulto_equivalente` no incluye una definición de "adulto equivalente". Quizás esta definición sea obvia entre sociólogos, pero para alguien como yo que no formar parte de esta área no me queda claro a qué se refiere esto. Creo que al no proveer definiciones un poco más detalladas se podría estar limitando a las personas que puedan encontrar el paquete útil.

**Problemas con ejemplos**
El ejemplo incluido bajo `map_agglomerates` me dio como resultado solo un mapa base. Al hacer una revisión más detallada, noté que el problema no era la base de datos `toybase_individual_2016_04` ni los cálculos realizados sobre ella, sino más bien en la función `map_agglomerates`. Al revisar el código para esta función noté que si removía `leaflet::addProviderTiles(leaflet::providers$Wikimedia) %>%` de la línea 12, obtengo el resultado esperado: un mapa con círculos que representan el valor del indicador escogido. Adicionalmente, noté que la paleta de colores no permite cambios porque está codificada (hard coded) en la línea 8 (`pal <- leaflet::colorNumeric(palette = "viridis", domain = df$indicator)`). Al reemplazar `palette = "viridis"` con `palette = palette` la paleta de colores cambia en el mapa.

El ejemplo incluido bajo `organize_caes` me dio la siguiente advertencia:

```
In organize_caes(base = bases) : Convirtiendo PP04B_COD a character
```

No estoy segura de si este resultado es esperado o no.

**Resultados de revisión automática**

```
library(pkgreviewr)
pkgreview_create(pkg_repo = "holatam/eph",
                 review_parent = "revision_eph")
devtools::check("revision_eph/eph")

── R CMD check results ────────────────────────────────────────────────────────────────────── eph 0.6.1 ────
Duration: 4m 18.5s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
   * checking re-building of vignette outputs ... [68s] OK
   * checking for non-standard things in the check directory ... OK
   * checking for detritus in the temp directory ... OK
   * DONE

   Status: OK

── R CMD check results ────────────────────────────────────────────────────────────────────── eph 0.6.1 ────
Duration: 4m 18.5s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```

```
devtools::test("revision_eph/eph/")
ℹ Testing eph
ℹ Testing eph
✔ | F W S  OK | Context
✔ |   1     6 | calculate_errors [1.1s]

══ Results ═════════════════════════════════════════════════════════════════════════════════════════════════

package ‘testthat’ was built under R version 4.2.3
```

Las advertencias fueron las siguientes:

```
Warning (test-calculate_errors.R:11:3): checkear shape ds
There was 1 warning in `dplyr::mutate()`.
ℹ In argument: `ds = calculate_errors(...)`.
Caused by warning:
! Using an external vector in selections was deprecated in tidyselect 1.1.0.
ℹ Please use `all_of()` or `any_of()` instead.
  # Was:
  data %>% select(measure)

  # Now:
  data %>% select(all_of(measure))
....

Warning (test-get_eahu.R:12:3): selec variables
There was 1 warning in `dplyr::mutate()`.
ℹ In argument: `microdata = purrr::pmap(...)`.
Caused by warning:
! Using an external vector in selections was deprecated in tidyselect 1.1.0.
ℹ Please use `all_of()` or `any_of()` instead.
  # Was:
  data %>% select(vars)

  # Now:
  data %>% select(all_of(vars))
....

Warning (test-is_in_github.R:1:1): (code run outside of `test_that()`)
package ‘curl’ was built under R version 4.2.3
```
