### 2023-09-11 #593 @maurolepore
@guadag12, @lidefi87, muchísimas gracias por su trabajo detallado y puntual.

En el proximo comentario voy a registrar formalmente la entrega de sus revisiones.

Próximos pasos:

1. @caropradier incorporará las revisiones y nos avisará cuando esté lista para su consideración.
2. Yo les pediré que indiquen si los cambios satisfacen sus expectativas o si tienen más sugerencias.
