+++ R/calculate_errors.R
@@ -69,7 +69,7 @@ calculate_errors <- function(value, codigo_aglo = "Total", periodo_eph = "2014.0
   # Operacion
   tabla_referencia <- eph::errores_muestrales %>%
     dplyr::filter(codigo == codigo_aglo & periodo == periodo_eph) %>%
-    dplyr::select(x, measure)
+    dplyr::select(all_of(c("x", measure)))
 
   find_closest <- function(y) {
     tabla_referencia[[measure]][which.min(abs(tabla_referencia[["x"]] - y))]
