+++ R/get_microdata.R
@@ -98,7 +98,7 @@ get_microdata <- function(year = 2018,
       dplyr::mutate(filter_col = purrr::map_lgl(error, ~ !is.null(.x))) %>%
       dplyr::filter(filter_col) %>%
       dplyr::mutate(error_message = purrr::map(error, ~ purrr::pluck(.x, 1))) %>%
-      dplyr::select(-microdata, -error) %>%
+      dplyr::select(-any_of(c("microdata", "error"))) %>%
       tidyr::unnest(cols = c(error_message))
 
     if (nrow(errors) > 0) {
@@ -109,7 +109,7 @@ get_microdata <- function(year = 2018,
     }
 
     df <- df %>%
-      dplyr::select(-error)
+      dplyr::select(-any_of(c("error")))
 
     if (!is.null(destfile) & !destfile_exists) {
       saveRDS(df, file = destfile)
@@ -121,6 +121,6 @@ get_microdata <- function(year = 2018,
   } else {
     df %>%
       tidyr::unnest(microdata) %>%
-      dplyr::select(-c(year, period, type))
+      dplyr::select(-any_of(c("year", "period", "type")))
   }
 }
