+++ R/get_eahu_internal.R
@@ -61,7 +61,7 @@ get_eahu_internal <- function(year = 2010,
 
     base <- base %>%
       dplyr::rename_all(toupper) %>%
-      dplyr::select(vars)
+      dplyr::select(all_of(vars))
 
     return(base)
   }
