### 2023-09-11 41ff4f1
further improvement of select and latin1 strings
