+++ R/diccionario_regiones.R
@@ -1,10 +1,10 @@
-#' Diccionario de regiones segun diseño de registro de la EPH
+#' Diccionario de regiones segun diseno de registro de la EPH
 #'
 #' @format  Un data frame con 6 filas y 2 variables
 #'
 #' \describe{
 #' \item{codigo}{Codigo de region}
-#' \item{region}{Nombre de región}
+#' \item{region}{Nombre de region}
 #'
 #' }
 "diccionario_regiones"
