+++ R/diccionario_aglomerados.R
@@ -1,4 +1,4 @@
-#' Diccionario de aglomerados segun diseño de registro de la EPH
+#' Diccionario de aglomerados segun diseno de registro de la EPH
 #'
 #' @format  Un data frame con 32 filas y 2 variables
 #'

[…]
