+++ R/adulto_equivalente.R
@@ -8,4 +8,10 @@
 #' \item{adequi}{unidades de adulto equivalente}
 #'
 #' }
+#'
+#' @details
+#' De acuerdo con la metodologia de INDEC (https://www.indec.gob.ar/ftp/cuadros/sociedad/preguntas_frecuentes_cba_cbt.pdf):
+#'
+#' "Para identificar la unidad consumidora de referencia, o adulto equivalente, se analizo en que grupo de edad se ubicaba la mayor concentracion de poblacion activa, y fue la categoria de hombres de 30 a 60 anios de edad la que presentaba el mayor peso relativo dentro del conjunto total de la poblacion de referencia. Por este motivo se considero como adulto equivalente al hombre de 30 a 60 anios con un estilo de vida moderadamente activo y no otro grupo poblacional."
+#'
 "adulto_equivalente"
