+++ man/canastas_reg_example.Rd
@@ -9,7 +9,7 @@ Un data frame con 12 filas y 5 variables
 
 \describe{
 \item{region}{Nombre de region}
-\item{periodo}{Año.trimestre}
+\item{periodo}{Anio.trimestre}
 \item{CBA}{canasta basica alimentaria}
 \item{CBT}{inversa del coeficiente de engels}
 \item{codigo}{codigo de region}
