+++ man/diccionario_regiones.Rd
@@ -3,13 +3,13 @@
 \docType{data}
 \name{diccionario_regiones}
 \alias{diccionario_regiones}
-\title{Diccionario de regiones segun diseño de registro de la EPH}
+\title{Diccionario de regiones segun diseno de registro de la EPH}
 \format{
 Un data frame con 6 filas y 2 variables
 
 \describe{
 \item{codigo}{Codigo de region}
-\item{region}{Nombre de región}
+\item{region}{Nombre de region}
 
 }
 }
@@ -17,6 +17,6 @@ Un data frame con 6 filas y 2 variables
 diccionario_regiones
 }
 \description{
-Diccionario de regiones segun diseño de registro de la EPH
+Diccionario de regiones segun diseno de registro de la EPH
 }
 \keyword{datasets}
