+++ man/calculate_errors.Rd
@@ -2,7 +2,7 @@
 % Please edit documentation in R/calculate_errors.R
 \name{calculate_errors}
 \alias{calculate_errors}
-\title{Calculo del desvío estándar y el coeficiente de variación}
+\title{Calculo del desvio estandar y el coeficiente de variacion}
 \usage{
 calculate_errors(
   value,
@@ -12,25 +12,25 @@ calculate_errors(
 )
 }
 \arguments{
-\item{value}{Vector numérico de las estimaciones de población para las
-que se desea hallar el desvío estándar o el coeficiente de variación.}
+\item{value}{Vector numerico de las estimaciones de poblacion para las
+que se desea hallar el desvio estandar o el coeficiente de variacion.}
 
-\item{codigo_aglo}{default = "Total". String con el código numerico del aglomerado al que
+\item{codigo_aglo}{default = "Total". String con el codigo numerico del aglomerado al que
 pertenecen las estimaciones. "Total" para trabajar estimaciones del conjunto de 31 aglomerados urbanos.}
 
 \item{periodo_eph}{default = "2014.03". String indicando el periodo al que corresponde la EPH. "2014.03" para
 obtener los errores muestrales correspondientes al tercer trimestre de 2014 en adelante. "2003.03_2014.02"
 para los errores muestrales del tercer trimestre del 2003 al segundo trimestre del 2014.}
 
-\item{measure}{default = "cv". String indicando la medida que se desea obtener. "cv" para obtener el coeficiente de variación
-correspondiente a las estimaciones o "ds" para obtener el desvío estándar.}
+\item{measure}{default = "cv". String indicando la medida que se desea obtener. "cv" para obtener el coeficiente de variacion
+correspondiente a las estimaciones o "ds" para obtener el desvio estandar.}
 }
 \value{
 Devuelve la estimacion de un total poblacional agregando el desvio estandar o el coeficiente de variacion correspondiente segun las tablas de error muestral de INDEC para EPH continua a partir del segundo trimestre 2003
 }
 \description{
-Asigna a una estimación de un total poblacional el desvío estándar o el coeficiente de variación
-correspondiente según las tablas de error muestral de INDEC para EPH continua
+Asigna a una estimacion de un total poblacional el desvio estandar o el coeficiente de variacion
+correspondiente segun las tablas de error muestral de INDEC para EPH continua
  a partir del segundo trimestre 2003.
 (Ver `errores_muestrales`)
 }
