+++ doc/eph.R
@@ -12,59 +12,75 @@ library(purrr)
 library(ggplot2)
 
 ## -----------------------------------------------------------------------------
-ind_3_18 <- get_microdata(year = 2018,
-                          trimester = 2,
-                          type = 'individual')
-
-## ----message=FALSE, warning=FALSE---------------------------------------------
-ind_2_02 <- get_microdata(year = 2001,
-                          wave = 2,
-                          type = 'individual')
+ind_3_18 <- get_microdata(
+  year = 2018,
+  trimester = 3,
+  type = "individual"
+)
 
 ## -----------------------------------------------------------------------------
-ind_3_18 <- organize_labels(df = ind_3_18,
-                            type='individual')
+ind_3_18 <- organize_labels(
+  df = ind_3_18,
+  type = "individual"
+)
 
 ## -----------------------------------------------------------------------------
-hog_3_18 <- get_microdata(year = 2018,
-                          trimester = 3,
-                          type='hogar') %>%
+hog_3_18 <- get_microdata(
+  year = 2018,
+  trimester = 3,
+  type = "hogar"
+) %>%
   organize_labels(.,
-                  type = 'hogar')
+    type = "hogar"
+  )
 
 ## -----------------------------------------------------------------------------
-calculate_tabulates(base = ind_3_18,
-                    x = 'NIVEL_ED',
-                    y = 'CH04',
-                    weights = 'PONDERA',
-                    add.totals = 'row',
-                    add.percentage='col')
+calculate_tabulates(
+  base = ind_3_18,
+  x = "NIVEL_ED",
+  y = "CH04",
+  weights = "PONDERA",
+  add.totals = "row",
+  add.percentage = "col"
+)
 
 ## -----------------------------------------------------------------------------
-calculate_tabulates(base = ind_3_18,
-                    x = 'NIVEL_ED',
-                    y = 'CH04',
-                    add.totals = 'row',
-                    add.percentage = 'col')
+calculate_tabulates(
+  base = ind_3_18,
+  x = "NIVEL_ED",
+  y = "CH04",
+  add.totals = "row",
+  add.percentage = "col"
+)
 
 ## -----------------------------------------------------------------------------
-### Armo vector con el nombre de las variables de interes
-variables <- c('CODUSU','NRO_HOGAR','COMPONENTE','ANO4',
-               'TRIMESTRE','CH04', 'CH06', #variables necesarias para hacer el panel
-               'ESTADO','PONDERA') #variables que nos interesan en nuestro analisis
+### Armo vector con el nombre de las variables de interés incluyendo
+# -variables necesarias para hacer el panel
+# -variables que nos interesan en nuestro análisis
+variables <- c(
+  "CODUSU", "NRO_HOGAR", "COMPONENTE", "ANO4",
+  "TRIMESTRE", "CH04", "CH06", 
+  "ESTADO", "PONDERA"
+) 
 
 ### Descargo la base individual para el 2018_t1
-base_2018t1 <- get_microdata(year = 2018,  trimester = 1,  type = 'individual',
-                       vars = variables)
+base_2018t1 <- get_microdata(
+  year = 2018, trimester = 1, type = "individual",
+  vars = variables
+)
 
 ### Descargo la base individual para el 2018_t2
-base_2018t2 <- get_microdata(year = 2018,  trimester = 2,  type = 'individual',
-                       vars = variables)
+base_2018t2 <- get_microdata(
+  year = 2018, trimester = 2, type = "individual",
+  vars = variables
+)
 
 ### Armo el panel
-pool <- organize_panels(bases = list(base_2018t1, base_2018t2),
[…]
