### 2023-09-11 #593 @caropradier
@guadag12 gracias de nuevo por tu trabajo!
Incorporé tus comentarios, incluyo algunas observaciones para seguir intercambiando en caso de que sea necesario:

1. Respecto a los Ejemplos que no corren localmente. En los tres casos, los ejemplos no corren localmente porque se trata de funciones nuevas (o que fueron modificadas) desde el último submit a CRAN. Si se carga el paquete usando devtools::[install_github](https://remotes.r-lib.org/reference/install_github.html)("holatam/eph"), los ejemplos funcionan normalmente (al completar el proceso de revisión pienso enviar a CRAN la última versión, por lo que este conflicto se resolverá).

2. Gracias por señalar la cuestión del elapsed time! Si no me equivoco, cuando estas dos funciones se incorporen a la versión en CRAN el problema dejará de ser de importancia (los tests incluyen skip_on_cran()).

3. Respecto a los tests, obviamente estoy de acuerdo con la pertinencia de testear lo máximo posible. Simplemente, la función package_coverage() no tiene en cuenta los tests en algunos casos (al usar skip_if_xxxxx()). Por ejemplo, get_microdata() aparece con 0% a pesar de contar con tres tests (que se encuentran en test-get_microdata.R).

4. Respecto a las líneas de código superan los 80 caracteres, coincido en que no es una buena práctica, únicamente sucede en scripts de input de datos, no en aquellos que describen las funciones, donde tenemos como prioridad que el código sea claro y fácil de leer.

5. Respecto a los carácteres Latin-1, creo haber eliminado los casos donde esto ocurría, gracias por señalarlo!

@lidefi87 gracias de nuevo por tus esfuerzos!
Incorporé también tus comentarios, incluyo algunas observaciones para seguir intercambiando en caso de que sea necesario:

1. Respecto a los Ejemplos que no corren localmente, entiendo que aplica lo mismo que señalé anteriormente para la primera revisión.

2. Respecto al nombre de las funciones, intentaría no hacer modificaciones mayores para no perturbar el flujo de trabajo de nuestros usuarios actuales (aproximadamente 30 mil personas usan el paquete y quisiera evitar generarles inconvenientes si no se trata de algo fundamental para el funcionamiento del paquete). No obstante, estoy de acuerdo con que se trata de una buena práctica, y lo tendré presente al incorporar nuevas funciones.

3. Muchas gracias por tu observación respecto a los acrónimos y otros conceptos tales como adulto_equivalente, no lo había notado y es cierto que puede ser confuso para nuevos usuarios. Incorporé modificaciones en la documentación para mejorar este aspecto.

4. Desarrollé mejor lo que sucede al aplicar organize_labels() en la viñeta y la documentación de la función, espero que ahora sea más claro!

5. Gracias por identificar el problema en map_agglomerates()! No lo había notado y ahora su funcionamiento es el esperado.

6. Respecto a organize_caes(), efectivamente ese es el comportamiento esperado. Simplemente se informa al usuario de esta transformación.

7. Modifiqué la sintaxis depreciada de select() por select( all_of() / any_of() )
