### 2023-09-12 #593 @maurolepore
@caropradier, gracias por incorporar los cambios pronto.

--

@guadag12 y @lidefi87, por favor indiquen una de estas dos opciones:

1. Recomiendan más cambios.
2. Aprueban el paquete. En tal caso, por favor, expresen su aprobación formalmente utilizando la [plantilla de aprobación de revisión](https://devdevguide.netlify.app/es/approval2template.es).
