### 2023-09-14 #593 @lidefi87
## Respuesta de quien hizo la revisión

#### Aprobación final (posterior a la revisión)

- [ x ] **Quien mantiene el paquete ha respondido a mi revisión y ha realizado cambios satisfactorios. Recomiendo aprobar este paquete.**

Estimación de horas dedicadas a la revisión: 0.5 horas

@caropradier gracias por hacer los cambios tan rápido. No tengo sugerencias adicionales, y creo que las que ofrecí ya han sido incluidas en la versión más reciente. Volví a correr las pruebas sugeridas en la revisión e incluyo los resultados abajo para su información.

En cuanto al punto 2 sobre el nombre de las funciones. Estoy totalmente de acuerdo contigo. Creo que no valdría la pena modificarlas ahora, pero sí creo que sería bueno tomar en cuenta el standard de ROpenSci en funciones futuras.

Nuevamente muy buen trabajo 😊

```
devtools::test("revision_eph/eph/")
── R CMD check results ───────────────────────────────────────────────────────────────────── eph 0.6.1 ────
Duration: 3m 25.1s

0 errors ✔ | 0 warnings ✔ | 0 notes ✔
```

```
devtools::check("revision_eph/eph")
══ Results ════════════════════════════════════════════════════════════════════════════════════════════════
Duration: 40.6 s

[ FAIL 0 | WARN 0 | SKIP 0 | PASS 57 ]
```
