### 2023-09-14 #593 @guadag12
## Respuesta de quien hizo la revisión

#### Aprobación final (posterior a la revisión)

- [ x] **Quien mantiene el paquete ha respondido a mi revisión y ha realizado cambios satisfactorios. Recomiendo aprobar este paquete.**

Estimación de horas dedicadas a la revisión: 0

Gracias por tan pronta respuesta, @caropradier! Excelente los cambios. De mi parte no hay más modificaciones,

Me encantó el trabajo que han hecho, felicitaciones! 🎉

Saludos!
