### 2023-09-14 #593 @yabellini
> @ropensci/blog-editors,
>
> eph puede ser de gran interés:
>
>     * Mucha gente ya usa este paquete.
>
>     * Según creo, es el primer paquete de rOpenSci que resultó del programa de campeonas y campeones.
>
>     * Fue revisado en Castellano.
>
>
> cc: @yabellini

Sii exacto! por todos los puntos que mencionas!. Que emoción ver esto hecho una realidad.  Felicitaciones a todes y muchas gracias  todas las personas involucradas.  Hablaré con Caro pero esto puede ser un artículo con todos los autores del paquete, no solo Caro, nuestra campeona.

Otra cosa mas, tenemos una etiqueta del programa de campeones para agregar en el repositorio del paquete.  Compartiré los detalles para que la puedan agregar.
