### 2023-10-23 #593 @caropradier
Hola! Perdón por la demora. Todo listo:

Por el momento dejaría la página de pkgdown si no es un problema, cuando tenga un poco más de tiempo puedo ocuparme de hacer esa migración :)
